ModbusTCP/requirements.txt
pymodbus
twisted
threading
RPi.GPIO
Adafruit_DHT
